Django/requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cycler==0.10.0
Django==3.1.5
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
joblib==1.0.0
kiwisolver==1.3.1
kombu==5.0.2
matplotlib==3.3.4
numpy==1.20.0
pandas==1.2.1
Pillow==8.1.0
prompt-toolkit==3.0.14
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.5
scikit-learn==0.24.1
scikit-multiflow==0.5.3
scipy==1.6.0
six==1.15.0
sortedcontainers==2.3.0
sqlparse==0.4.1
threadpoolctl==2.1.0
vine==5.0.0
wcwidth==0.2.5
